pyproject.toml
[project]
name = "algebraics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dynaconf>=3.2.10",
    "glfw>=2.8.0",
    "imageio>=2.37.0",
    "matplotlib>=3.10.1",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.1",
    "pygame>=2.6.1",
    "pyopengl>=3.1.9",
    "pyopengl-accelerate>=3.1.9",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]

[build-system]
    requires = ["poetry-core"]
    build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.select = ["I"]
line-length = 88
